Improved CfgColor

+++ config/config.h
@@ -12,7 +12,7 @@ typedef struct {
     uint8_t r;
     uint8_t g;
     uint8_t b;
-    float a;
+    uint8_t a;
 } CfgColor;
 
 typedef enum {
